+++ CHANGELOG.md
@@ -45,6 +45,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Minimap compass
 - Initial crafting system implementation
 - Protection stat to armor that reduces incoming damage
+- Loading-Screen tips
 
 ### Changed